requirements.txt
streamlit==1.22.0
openai==0.27.8
python-dotenv==1.0.0